Crate moore_vhdl

This crate implements VHDL for the moore compiler.

Re-exports

pub extern crate moore_common;

pub extern crate moore_vhdl_syntax as syntax;

pub use moore_common as common;

Modules

add_ctx
A context within which nodes can be added.

arenas
Multi-type arena allocation

builtin
Builtin libraries, packages, types, and functions.

codegen
LLHD code generation for VHDL.

debug

defs
A compiler pass that gathers definitions.

hir
The High-level Intermediate Representation of a VHDL design.

konst
This module implements constant value calculation for VHDL.

konst2
This module implements constant values for VHDL.

lazy
An implementation of lazy compiler passes.

make_ctx
A context within which compiler passes can be described.

op
Operators

overload_resolver
Overload resolution for subprograms and enum literals.

scope
Facilities to manage declarations and resolve names.

scope2
Facilities to manage declarations and resolve names.

score
This module implements the scoreboard that drives the compilation of VHDL.

symtbl

term
Expressions

ty
This module implements VHDL types.

ty2
The VHDL type system.

typeck
This module implements the type calculation of the scoreboard.

Macros

make_arenas
Generate a collection of arenas for different types.