Struct moore_vhdl::add_ctx::AddContext

pub struct AddContext<'sbc, 'lazy: 'sbc, 'sb: 'lazy, 'ast: 'sb, 'ctx: 'sb> {
    pub ctx: &'sbc ScoreContext<'lazy, 'sb, 'ast, 'ctx>,
    pub scope: ScopeRef,
}

A context within which nodes can be added.

See the module documentation for details.

Fields
ctx: &'sbc ScoreContext<'lazy, 'sb, 'ast, 'ctx>

The outer context.

scope: ScopeRef

The scope to which items will be added.

Methods

impl<'sbc, 'lazy, 'sb, 'ast, 'ctx> AddContext<'sbc, 'lazy, 'sb, 'ast, 'ctx>

pub fn new(
ctx: &'sbc ScoreContext<'lazy, 'sb, 'ast, 'ctx>,
scope: ScopeRef
) -> AddContext<'sbc, 'lazy, 'sb, 'ast, 'ctx>

Create a new context.

pub fn with_scope(
&self,
scope: ScopeRef
) -> AddContext<'sbc, 'lazy, 'sb, 'ast, 'ctx>

Create a new context with different scope.

pub fn make<I>(
&self,
span: Span
) -> (MakeContext<'sbc, 'lazy, 'sb, 'ast, 'ctx, I>, I, ScopeRef) where
I: NodeRef,

Create a new context for describing a node.

pub fn unimp<T, R>(&self, node: &T) -> Result<R> where
T: HasSpan + HasDesc,

Emit a diagnostic that a node is not implemented.

pub fn add_optional<T, F, R>(
&self,
node: &'ast Option<T>,
f: F
) -> Result<Option<R>> where
F: FnOnce(&Self, &'ast T) -> Result<R>,

Add an optional node.

This convenience function performs a few things at a time. First it applies a closure f to the value of an option, returning an error if the closure returns one. Otherwise it wraps the result in an option and returns it.

impl<'sbc, 'lazy, 'sb, 'ast, 'ctx> AddContext<'sbc, 'lazy, 'sb, 'ast, 'ctx>

pub fn add_subtype_ind(&self, ind: &'ast SubtypeInd) -> Result<SubtypeIndRef>

Add a subtype indication.

pub fn add_subtype_ind_hir(&self, hir: SubtypeInd) -> Result<SubtypeIndRef>

Add a subtype indication already lowered to HIR.

pub fn schedule_subtype_ind(&self, mk: &MakeContext<SubtypeIndRef>)

Schedule subtype indication tasks.

impl<'sbc, 'lazy, 'sb, 'ast, 'ctx> AddContext<'sbc, 'lazy, 'sb, 'ast, 'ctx>

pub fn add_expr(&self, expr: &'ast Expr) -> Result<ExprRef>

Add an expression.

pub fn add_expr_hir(&self, hir: Expr) -> Result<ExprRef>

Add an expression already lowered to HIR.

pub fn schedule_expr(&self, mk: &MakeContext<ExprRef>)

Schedule expression tasks.

pub fn add_choices<I>(&self, ast: Spanned<I>) -> Result<Spanned<Choices>> where
I: IntoIterator<Item = &'ast Expr>,

Add a list of choices.

pub fn add_discrete_range(
&self,
ast: &'ast Expr
) -> Result<Spanned<DiscreteRange>>

Add a discrete range.

pub fn add_aggregate_hir(&self, hir: Aggregate) -> Result<AggregateRef>

Add an aggregate already lowered to HIR.

pub fn schedule_aggregate(&self, mk: &MakeContext<AggregateRef>)

Schedule aggregate tasks.

impl<'sbc, 'lazy, 'sb, 'ast, 'ctx> AddContext<'sbc, 'lazy, 'sb, 'ast, 'ctx>

pub fn add_seq_stmts<I, B>(&self, stmts: I, container_name: &str) -> Result<B> where
I: IntoIterator<Item = &'ast Stmt>,
B: FromIterator<SeqStmtRef>,

Add multiple sequential statements.

pub fn add_seq_stmt(
&self,
stmt: &'ast Stmt,
container_name: &str
) -> Result<SeqStmtRef>

Add a sequential statement.

pub fn add_wait_stmt(&self, stmt: &'ast Stmt) -> Result<WaitStmtRef>

Add a wait statement.

pub fn add_assert_stmt(&self, stmt: &'ast Stmt) -> Result<AssertStmtRef>

Add an assert statement.

pub fn add_report_stmt(&self, stmt: &'ast Stmt) -> Result<ReportStmtRef>

Add a report statement.

pub fn add_sig_assign_stmt(&self, stmt: &'ast Stmt) -> Result<SigAssignStmtRef>

Add a sig_assign statement.

pub fn add_var_assign_stmt(&self, stmt: &'ast Stmt) -> Result<VarAssignStmtRef>

Add a var_assign statement.

pub fn add_call_stmt(&self, stmt: &'ast Stmt) -> Result<CallStmtRef>

Add a call statement.

pub fn add_if_stmt(&self, stmt: &'ast Stmt) -> Result<IfStmtRef>

Add an if statement.

pub fn add_case_stmt(&self, stmt: &'ast Stmt) -> Result<CaseStmtRef>

Add a case statement.

pub fn add_loop_stmt(&self, stmt: &'ast Stmt) -> Result<LoopStmtRef>

Add a while loop statement.

pub fn add_nexit_stmt(&self, stmt: &'ast Stmt) -> Result<NexitStmtRef>

Add a next statement.

pub fn add_return_stmt(&self, stmt: &'ast Stmt) -> Result<ReturnStmtRef>

Add a return statement.

pub fn add_null_stmt(&self, stmt: &'ast Stmt) -> Result<NullStmtRef>

Add a null statement.

pub fn add_sensitivity_list<I>(
&self,
ast: Spanned<I>
) -> Result<Spanned<SensitivityList>> where
I: IntoIterator<Item = &'ast CompoundName>,

Add a sensitivity list.

pub fn add_label(&self, ast: &'ast Spanned<Name>) -> Result<Spanned<StmtRef>>

Add a label.

Immediately resolves the label name and returns the corresponding statement.

pub fn add_target(
&self,
ast: &'ast Spanned<AssignTarget>
) -> Result<Spanned<Target>>

Add a target variable.

This is an target for a variable assignment.

impl<'sbc, 'lazy, 'sb, 'ast, 'ctx> AddContext<'sbc, 'lazy, 'sb, 'ast, 'ctx>

pub fn add_const_decl<I>(&self, decl: &'ast ObjDecl) -> Result<Vec<I>> where
I: From<ConstDeclRef>,

Add a constant declaration.

pub fn add_signal_decl<I>(&self, decl: &'ast ObjDecl) -> Result<Vec<I>> where
I: From<SignalDeclRef>,

Add a signal declaration.

pub fn add_var_decl<I>(&self, decl: &'ast ObjDecl) -> Result<Vec<I>> where
I: From<VarDeclRef>,

Add a variable declaration.

pub fn add_file_decl<I>(&self, decl: &'ast ObjDecl) -> Result<Vec<I>> where
I: From<FileDeclRef>,

Add a file declaration.

impl<'sbc, 'lazy, 'sb, 'ast, 'ctx> AddContext<'sbc, 'lazy, 'sb, 'ast, 'ctx>

pub fn add_type_decl(&self, decl: &'ast TypeDecl) -> Result<TypeDeclRef>

Add a type declaration.

pub fn add_type_data(
&self,
id: TypeDeclRef,
name: Spanned<Name>,
data: &'ast Spanned<TypeData>
) -> Result<Spanned<TypeData>>

Add a type definition.