Module moore_vhdl::hir

The High-level Intermediate Representation of a VHDL design.

Modules

visit
Visitor pattern for the HIR.

Structs

Aggregate
An aggregate.

AllocContext
A context for HIR node construction.

Arch

Arenas

Arenas2
An arena to allocate HIR nodes into.

ArrayConstraint
An array constraint.

AssertStmt
An assertion statement.

AssocElement
An association element.

CallStmt
A procedure call statement.

CaseStmt
A case statement.

Cond
A conditional waveform or expression.

ConstDecl
A constant declaration.

Decl
An object declaration.

Entity

Expr
An expression.

FileDecl
A file declaration.

IfStmt
An if statement.

IntfSignal

Lib

Library
A library.

LitExpr
A literal expression.

LoopStmt
A loop statement.

NexitStmt
A next or exit statement.

NullStmt
A null statement.

Package
A package declaration.

Package2

PackageBody
A package body.

PackageInst
A package instantiation.

ProcessStmt
A process statement.

RecordConstraint
A record constraint as per IEEE 1076-2008 section 5.3.3.

ReportStmt
A report statement.

ReturnStmt
A return statement.

Sel
A selected waveform or expression.

SigAssignStmt
A sequential signal assignment.

SignalDecl
A signal declaration.

Slot
A placeholder for an HIR node.

Stmt
A statement.

Subprog
A subprogram.

SubprogBody
A subprogram body.

SubprogInst
A subprogram instantiation.

SubprogSpec
A subprogram specification.

SubtypeDecl
A subtype declaration as per IEEE 1076-2008 section 6.3.

SubtypeDecl2
A subtype declaration.

SubtypeInd

SubtypeInd2
A subtype indication.

TypeDecl

TypeDecl2
A type declaration.

VarAssignStmt
A variable assignment statement.

VarDecl
A variable declaration.

WaitStmt
A wait statement.

WaveElem
An element of a signal assignment waveform.

Enums

AggregateKind
An aggregate kind.

ArrayChoice
A choice in an array aggregate.

ArrayTypeIndex
An index of an array type.

AssocActual
An actual part of an association element.

Choice
A choice in an aggregate.

Constraint
A constraint.

DelayMechanism
The delay mechanism of a normal signal assignment.

Dir

DiscreteRange
A discrete range.

ElementConstraint
An element constraint.

EnumLit
An enumeration literal as listed in a type declaration.

ExprData
An expression variant.

ForceMode
The mode of a signal force/release statement.

IntfSignalMode

LitExprValue
The value of a literal expression.

LoopScheme
A loop scheme.

NexitMode
A discriminant for next/exit statements.

ProcessSensitivity
A process sensitivity specification.

Range
A range.

Range2
A range.

SigAssignKind
A signal assignment kind.

SigAssignTarget
A signal assignment target.

SignalKind
A signal kind.

SubprogKind
A subprogram kind.

Target
A target.

TypeData
The meat of a type declaration.

VarAssignKind
A variable assignment kind.

Traits

Decl2
Common functions of declaration HIR node.

Expr2
An expression.

ExprContext
A context that provides the facilities to operate on expressions.

FromAst
Construct something from an AST node.

LatentNode
Lazily resolve to a Node.

Node
Common functions of HIR nodes.

Functions

apply_use_clauses

Type Definitions

ArrayChoices
A list of choices used in array aggregates.

AssocList
An association list.

Choices
A list of choices used in aggregates, selected assignments, and case statements.

RecordChoices
A list of choices used in record aggregates.

SensitivityList
A sensitivity list.

UnitTable
A table of units for a physical type.

Waveform
A signal assignment waveform.