Module moore_vhdl::score

This module implements the scoreboard that drives the compilation of VHDL.

Re-exports

pub use crate::builtin::*;

Structs

AggregateRef

AliasDeclRef

ArchRef

ArchTable
A table of the architectures in a library, and how they relate to the entities.

Arenas
A collection of arenas that the scoreboard uses to allocate its nodes.

ArrayTypeIndexRef

AssertStmtRef

AstTable

AttrDeclRef

AttrSpecRef

BlockStmtRef

BuiltinOpRef

BuiltinPkgRef

CallStmtRef

CaseGenStmtRef

CaseStmtRef

CfgRef

CfgSpecRef

CompDeclRef

CompInstStmtRef

ConcAssertStmtRef

ConcCallStmtRef

ConcSigAssignStmtRef

ConstDeclRef

CtxItemsRef

CtxRef

DeclValueRef

DefValueRef

DesignUnitRef

DisconSpecRef

EntityArchTable
A table of the architectures associated with an entity.

EntityRef

EnumRef
A reference to an enumeration literal, expressed as the type declaration which defines the enumeration and the index of the literal.

ExprRef

FileDeclRef

ForGenStmtRef

GenericMapRef

GroupDeclRef

GroupTempRef

HirTable

IfGenStmtRef

IfStmtRef

IntfConstRef

IntfFileRef

IntfPkgRef

IntfSignalRef

IntfSubprogRef

IntfTypeRef

IntfVarRef

LatentPkgRef

LatentSubprogRef

LatentTypeMarkRef

LibRef

LoopStmtRef

NexitStmtRef

NullStmtRef

PkgBodyRef

PkgDeclRef

PkgInstRef

PortMapRef

ProcessStmtRef

ReportStmtRef

ReturnStmtRef

Scope
A scope.

ScoreBoard
The VHDL scoreboard that keeps track of compilation results.

ScoreContext
The VHDL context which holds information about the language scoreboard and the global scoreboard in its language-agnostic generic form. All useful operations are defined on this context rather than on the scoreboard directly, to decouple processing and ownership.

SigAssignStmtRef

SignalDeclRef

SubprogBodyRef

SubprogDeclRef

SubprogInstRef

SubtypeDeclRef

SubtypeIndRef

TypeDeclRef

UnitRef
A reference to a physical unit, expressed as the type declaration which defines the unit and the index of it.

VarAssignStmtRef

VarDeclRef

WaitStmtRef

Enums

ConcStmtRef

DeclInBlockRef

DeclInPkgBodyRef

DeclInPkgRef

DeclInProcRef

DeclInSubprogRef

Def

GenericRef

IntfObjRef

LatentName
A general name in the AST that can be resolved. Used for e.g. for package and subprogram bodies to resolve the name of their target.

PkgRef

ResolvableName
A name that can be resolved in a scope.

ScopeRef

SeqStmtRef

SignalRef

StmtRef

SubprogRef

TypeCtx
The type requirements imposed upon an expression by its context. This is needed for overload resolution, where the type of the overload to be picked is determined by the context in which the expression appears.

TypeMarkRef

TypedNodeRef

Type Definitions

Defs
A set of names and definitions.