Struct moore_vhdl::score::ScoreContext

pub struct ScoreContext<'lazy, 'sb: 'lazy, 'ast: 'sb, 'ctx: 'sb> {
    pub sess: &'lazy Session,
    pub global: &'lazy dyn GenericContext,
    pub sb: &'sb ScoreBoard<'ast, 'ctx>,
    pub lazy: &'lazy LazyPhaseTable<'sb, 'ast, 'ctx>,
}

The VHDL context which holds information about the language scoreboard and the global scoreboard in its language-agnostic generic form. All useful operations are defined on this context rather than on the scoreboard directly, to decouple processing and ownership.

Fields
sess: &'lazy Session

The compiler session which carries the options and is used to emit diagnostics.

global: &'lazy dyn GenericContext

The global context.

sb: &'sb ScoreBoard<'ast, 'ctx>

The VHDL scoreboard.

lazy: &'lazy LazyPhaseTable<'sb, 'ast, 'ctx>

The table of scheduled operations.

Methods

impl<'lazy, 'sb, 'ast, 'ctx> ScoreContext<'lazy, 'sb, 'ast, 'ctx>

pub fn unpack_expr(
&self,
ast: &'ast Expr,
scope_id: ScopeRef
) -> Result<ExprRef>

Unpack an AST expression.

pub fn unpack_subtype_ind(
&self,
ast: &'ast SubtypeInd,
scope_id: ScopeRef
) -> Result<SubtypeIndRef>

Unpack an AST subtype indication.

pub fn unpack_type_mark(
&self,
ast: LatentName<'ast>,
scope_id: ScopeRef
) -> Result<Spanned<LatentTypeMarkRef>>

Unpack a compound name as a type mark.

pub fn unpack_package_name(
&self,
ast: LatentName<'ast>,
scope_id: ScopeRef
) -> Result<Spanned<LatentPkgRef>>

Unpack a compound name as a package name.

pub fn unpack_subprog_name(
&self,
ast: LatentName<'ast>,
scope_id: ScopeRef
) -> Result<Spanned<LatentSubprogRef>>

Unpack a compound name as a subprogram name.

pub fn unpack_block_decls(
&self,
scope_id: ScopeRef,
decls: &'ast [DeclItem],
container_name: &str
) -> Result<Vec<DeclInBlockRef>>

Unpack a slice of AST declarative items into a list of items admissible in the declarative part of a block.

See IEEE 1076-2008 section 3.3.2.

pub fn unpack_process_decls(
&self,
scope_id: ScopeRef,
decls: &'ast [DeclItem],
container_name: &str
) -> Result<Vec<DeclInProcRef>>

Unpack a slice of AST declarative items into a list of items admissible in the declarative part of a process.

See IEEE 1076-2008 section 11.3.

pub fn unpack_subprog_decls(
&self,
scope_id: ScopeRef,
decls: &'ast [DeclItem]
) -> Result<Vec<DeclInSubprogRef>>

Unpack a slice of AST declarative items into a list of items admissible in the declarative part of a subprogram.

See IEEE 1076-2008 section 4.3.

pub fn unpack_concurrent_stmts(
&self,
scope_id: ScopeRef,
stmts: &'ast [Stmt],
container_name: &str
) -> Result<Vec<ConcStmtRef>>

Unpack a slice of concurrent statements.

See IEEE 1076-2008 section 11.1.

pub fn unpack_sequential_stmts(
&self,
scope_id: ScopeRef,
stmts: &'ast [Stmt],
container_name: &str
) -> Result<Vec<SeqStmtRef>>

Unpack a slice of sequential statements.

See IEEE 1076-2008 section 10.

pub fn unpack_signal_assign_target(
&self,
scope_id: ScopeRef,
target: &'ast AssignTarget
) -> Result<SigAssignTarget>

Unpack a signal assignment target.

See IEEE 1076-2008 section 10.5.2.1.

pub fn unpack_signal_assign_mode(
&self,
scope_id: ScopeRef,
mode: &'ast Spanned<AssignMode>,
tyctx: &TypeCtx<'ctx>
) -> Result<Spanned<SigAssignKind>>

Unpack a signal assignment mode.

See IEEE 1076-2008 section 10.5.

pub fn unpack_delay_mechanism(
&self,
scope_id: ScopeRef,
dm: &'ast Option<Spanned<DelayMech>>
) -> Result<DelayMechanism>

Unpack a delay mechanism.

See IEEE 1076-2008 section 10.5.2.1. If no mechanism is specified, inertial is assumed. Theoretically, the inertial transport mechanism is mapped to reject-inertial with the pulse rejection limit determined by the delay of the first element in the waveform. We don't have that information readily available at this time, so we simply map to inertial and leave the resolution of this to stages further down the pipeline.

pub fn unpack_cond_waveforms(
&self,
_scope_id: ScopeRef,
waves: &'ast [CondWave],
_tyctx: &TypeCtx<'ctx>
) -> Result<Cond<Waveform>>

Unpack the the waves of a simple wave assignment.

pub fn unpack_waveform(
&self,
scope_id: ScopeRef,
wave: &'ast Wave,
tyctx: &TypeCtx<'ctx>
) -> Result<Waveform>

Unpack a single waveform.

See IEEE 1076-2008 section 10.5.2.

pub fn sanitize_paren_elems_as_exprs(
&self,
elems: &'ast [ParenElem],
context: &str
) -> Vec<&'ast Expr>

Ensure that parenthesis contain only a list of expressions.

This is useful since the parser generally expects parenthesized expressions of the form (expr|expr|expr => expr, expr) even in palces where only (expr, expr) would be applicable. This function takes the parenthesized expression and ensures it is of the latter form.

pub fn lower_subprog_spec(
&self,
scope_id: ScopeRef,
ast: &'ast SubprogSpec
) -> Result<SubprogSpec>

Lower an AST subprogram specification to HIR.

pub fn lower_subprog_name(
&self,
kind: SubprogKind,
name: &'ast PrimaryName
) -> Result<Spanned<ResolvableName>>

Lower the name of an AST subprogram to HIR and perform checks.

pub fn unpack_generics(
&self,
scope_id: ScopeRef,
decls: &'ast [IntfDecl],
into: &mut Vec<GenericRef>
) -> Result<()>

Unpack generics from a list of interface declarations.

See IEEE 1076-2008 section 6.5.6.1.

pub fn unpack_generic_map(
&self,
_scope_id: ScopeRef,
elems: &'ast ParenElems
) -> Result<Vec<()>>

Unpack a generic map from a parenthesized list of elements.

See IEEE 1076-2008 section 6.5.7.2.

impl<'lazy, 'sb, 'ast, 'ctx> ScoreContext<'lazy, 'sb, 'ast, 'ctx>

pub fn make_ctx_items_scope(
&self,
id: CtxItemsRef,
parent: Option<ScopeRef>
) -> Result<CtxItemsRef>

impl<'lazy, 'sb, 'ast, 'ctx> ScoreContext<'lazy, 'sb, 'ast, 'ctx>

pub fn add_library(&self, name: Name, id: LibRef, lib: Vec<&'ast DesignUnit>)

Add a library of AST nodes. This function is called by the global scoreboard to add VHDL-specific AST nodes.

pub fn span<I>(&self, id: I) -> Option<Span> where
I: Into<NodeId>,

Obtain the span associated with a node ID.

pub fn set_span<I>(&self, id: I, span: Span) where
I: Into<NodeId>,

Associate a span with a node ID.

pub fn bug<I>(&self, id: I, msg: String) where
I: Into<NodeId>,

Report a compiler bug associate with a node.

pub fn ast<I>(&self, id: I) -> <AstTable<'ast> as NodeStorage<I>>::Node where
I: 'ast + Copy + Debug,
AstTable<'ast>: NodeStorage<I>,
<AstTable<'ast> as NodeStorage<I>>::Node: Copy + Debug,

Obtain the AST node corresponding to a node reference. The AST node must have previously been added to the ast_table, otherwise this function panics.

pub fn set_ast<I>(&self, id: I, ast: <AstTable<'ast> as NodeStorage<I>>::Node) where
I: Copy + Debug,
AstTable<'ast>: NodeStorage<I>,

Store an AST node in the scoreboard.

pub fn hir<I>(&self, id: I) -> Result<<HirTable<'ctx> as NodeStorage<I>>::Node> where
I: 'ctx + Copy + Debug,
HirTable<'ctx>: NodeStorage<I>,
ScoreContext<'lazy, 'sb, 'ast, 'ctx>: NodeMaker<I, <HirTable<'ctx> as NodeStorage<I>>::Node>,
<HirTable<'ctx> as NodeStorage<I>>::Node: Copy + Debug,

Obtain the HIR of a node, generating it if needed. Returns an error if the HIR cannot be generated.

pub fn set_hir<I>(&self, id: I, hir: <HirTable<'ctx> as NodeStorage<I>>::Node) where
I: Copy + Debug,
HirTable<'ctx>: NodeStorage<I>,

Store the HIR of a node.

pub fn existing_hir<I>(
&self,
id: I
) -> Result<<HirTable<'ctx> as NodeStorage<I>>::Node> where
I: Copy + Debug,
HirTable<'ctx>: NodeStorage<I>,
<HirTable<'ctx> as NodeStorage<I>>::Node: Copy + Debug,

Obtain the HIR of a node. Returns an error if none exists.

pub fn lazy_hir<I, R>(&self, id: I) -> Result<&'ctx R> where
I: Copy + Debug,
R: Debug + 'ctx,
LazyHirTable<'sb, 'ast, 'ctx>: NodeStorage<I, Node = LazyNode<Box<dyn Fn(&'a ScoreContext<'b, 'sb, 'ast, 'ctx>) -> Result<R> + 'sb>>>,
HirTable<'ctx>: NodeStorage<I, Node = &'ctx R>,
Arenas: Alloc<'ctx, 'ctx, R>,

Determine the HIR for a node.

If the HIR is already in the scoreboard, returns it immediately. Otherwise the lazy HIR table is triggered which will compute the HIR via the corresponding closure.

pub fn defs(&self, id: ScopeRef) -> Result<&'ctx Defs>

pub fn archs(&self, id: LibRef) -> Result<&'ctx ArchTable>

pub fn lldecl<I>(&self, id: I) -> Result<ValueRef> where
I: 'ctx + Copy + Debug + Into<NodeId>,
ScoreContext<'lazy, 'sb, 'ast, 'ctx>: NodeMaker<I, DeclValueRef>,

pub fn lldef<I>(&self, id: I) -> Result<ValueRef> where
I: 'ctx + Copy + Debug + Into<NodeId>,
ScoreContext<'lazy, 'sb, 'ast, 'ctx>: NodeMaker<I, DefValueRef>,

pub fn ty<I>(&self, id: I) -> Result<&'ctx Ty> where
I: 'ctx + Copy + Debug + Into<NodeId>,
ScoreContext<'lazy, 'sb, 'ast, 'ctx>: NodeMaker<I, &'ctx Ty>,

Determine the type of a node.

If called for the first time with the given id, calculates the type by calling self.make(id). Otherwise returns the existing information.

pub fn lazy_typeck<I>(&self, id: I) -> Result<()> where
I: Into<NodeId>,

Check the type of a node.

If the node already had its type checked, immediately returns the result of that operation. Otherwise runs the task scheduled in the lazy table.

pub fn lazy_typeval<I>(&self, id: I) -> Result<&'ctx Ty> where
I: Into<NodeId>,

Determine the type of a node.

If the node already had its type determined, immediately returns the result of that operation. Otherwise runs the task scheduled in the lazy table.

pub fn scope(&self, id: ScopeRef) -> Result<&'ctx Scope>

pub fn const_value<I>(&self, id: I) -> Result<&'ctx Const> where
I: 'ctx + Copy + Debug + Into<NodeId>,
ScoreContext<'lazy, 'sb, 'ast, 'ctx>: NodeMaker<I, &'ctx Const>,

pub fn type_context<I>(&self, id: I) -> Option<TypeCtx<'ctx>> where
I: Copy + Debug + Into<NodeId>,

Obtain the type context for an expression.

Returns None if no context information is available.

pub fn type_context_resolved<I>(&self, id: I) -> Result<Option<&'ctx Ty>> where
I: Copy + Debug + Into<NodeId>,

Obtain the type indicated by the type context for an expression.

Returns None if no context information is available.

pub fn set_type_context<I, T>(&self, id: I, tyctx: T) where
I: Copy + Debug + Into<NodeId>,
TypeCtx<'ctx>: From<T>,

Store a type context for an expression.

Upon type checking, the expression is likely to consult this context to determine its type.

pub fn set_type_context_optional<I, T>(&self, id: Option<I>, tyctx: T) where
I: Copy + Debug + Into<NodeId>,
TypeCtx<'ctx>: From<T>,

Store a type context for an optional expression.

Upon type checking, the expression is likely to consult this context to determine its type. Does nothing if id.is_none().

impl<'lazy, 'sb, 'ast, 'ctx> ScoreContext<'lazy, 'sb, 'ast, 'ctx>

pub fn resolvable_from_primary_name(
&self,
primary: &PrimaryName
) -> Result<Spanned<ResolvableName>>

Convert a primary name as it is present in the AST to a resolvable name that can be defined and resolved in a scope.

pub fn resolve_name(
&self,
name: Spanned<ResolvableName>,
scope_id: ScopeRef,
only_defs: bool,
allow_fail: bool
) -> Result<Vec<Spanned<Def>>>

Resolve a name within a scope. Traverses to the parent scopes if nothing matching the name is found.

pub fn resolve_compound_name<'a>(
&self,
name: &'a CompoundName,
scope_id: ScopeRef,
only_defs: bool
) -> Result<(ResolvableName, Vec<Spanned<Def>>, Span, &'a [NamePart])>

Resolve a compound name within a scope.

pub fn builtin_boolean_type(&self) -> &'ctx Ty

Get the builtin type standard.boolean.

pub fn builtin_time_type(&self) -> &'ctx Ty

Get the builtin type standard.time.

pub fn builtin_string_type(&self) -> &'ctx Ty

Get the builtin type standard.string.

pub fn builtin_severity_type(&self) -> &'ctx Ty

Get the builtin type standard.severity.

impl<'lazy, 'sb, 'ast, 'ctx> ScoreContext<'lazy, 'sb, 'ast, 'ctx>

pub fn default_value_for_type(&self, ty: &Ty) -> Result<&'ctx Const>

Calculate the implicit default value for a type.

pub fn intern_const<T>(&self, konst: T) -> &'ctx Const where
T: Into<Const>,

Internalize a constant.

See ScoreBoard::intern_const.

pub fn intern_ty<T>(&self, ty: T) -> &'ctx Ty where
T: Into<Ty>,

Internalize a type.

See ScoreBoard::intern_ty.

impl<'lazy, 'sb, 'ast, 'ctx> ScoreContext<'lazy, 'sb, 'ast, 'ctx>

pub fn map_type(&self, ty: &Ty) -> Result<Type>

Map a VHDL type to the corresponding LLHD type.

pub fn map_const(&self, konst: &Const) -> Result<ValueRef>

Map a constant value to the LLHD counterpart.

impl<'lazy, 'sb, 'ast, 'ctx> ScoreContext<'lazy, 'sb, 'ast, 'ctx>

pub fn deref_named_type<'a>(&self, ty: &'a Ty) -> Result<&'a Ty> where
'ctx: 'a,

Replace Ty::Named by the actual type definition recursively.

impl<'lazy, 'sb, 'ast, 'ctx> ScoreContext<'lazy, 'sb, 'ast, 'ctx>

pub fn make_range_ty(
&self,
dir: Dir,
lb_id: ExprRef,
rb_id: ExprRef,
span: Span
) -> Result<&'ctx Ty>

impl<'lazy, 'sb, 'ast, 'ctx> ScoreContext<'lazy, 'sb, 'ast, 'ctx>

pub fn with_scope<F, R>(&self, scope: ScopeRef, f: F) -> Result<R> where
F: FnOnce(&mut Scope) -> Result<R>,

Lookup a scope and perform an operation on it.

pub fn subscope(&self, scope: ScopeRef, parent: ScopeRef)

Create a subscope of another scope.

pub fn define(
&self,
scope: ScopeRef,
name: Spanned<ResolvableName>,
def: Def
) -> Result<()>

Define a new name in a scope.

pub fn import_def(
&self,
scope: ScopeRef,
name: Spanned<ResolvableName>,
def: Def
) -> Result<()>

Import a definition into a scope.

pub fn import_scope(&self, scope: ScopeRef, into: ScopeRef) -> Result<()>

Import an entire scope into another scope.