Trait moore_vhdl::term::DefSpecificTermContext

pub trait DefSpecificTermContext<'t, D> {
    fn termify_defs(
        &self, 
        name: Spanned<ResolvableName>, 
        defs: Vec<Spanned<D>>
    ) -> Result<Spanned<Term<'t>>>;
}

Required methods

fn termify_defs(
&self,
name: Spanned<ResolvableName>,
defs: Vec<Spanned<D>>
) -> Result<Spanned<Term<'t>>>

Termify the result of a name resolution.

This ensures that there is only one definition, or in case of something overloadable, that all definitions are of the same kind.

Loading content...