Module moore_vhdl::ty

This module implements VHDL types.

Structs

ArrayTy
An array type.

DisplayList

EnumTy
An enumeration type. Rather than keeping track of each enumeration value in here, we simply point at the type declaration.

IntTy
An integer type.

PhysicalTy
A physical type.

PhysicalUnit
A unit of a physical type.

RecordTy
A record type.

SubprogTy
A subprogram type.

SubprogTyArg
A subprogram argument type.

Enums

ArrayIndex
An index type of an array type.

Dir

Ty

TyName
A type name.