Enum moore_vhdl::ty::Ty

pub enum Ty {
    Named(TyName, TypeMarkRef),
    Null,
    Int(IntTy),
    UniversalInt,
    UnboundedInt,
    Enum(EnumTy),
    Physical(PhysicalTy),
    Access(Box<Ty>),
    Array(ArrayTy),
    File(Box<Ty>),
    Record(RecordTy),
    Subprog(SubprogTy),
}

Variants

Named(TyName, TypeMarkRef)

A named type. In a signal declaration for example, the source code mentions the type of the signal. This type name is resolved to its actual declaration somewhere else in the source code. Thus this type acts as a sort of "pointer" to a type, together with information on how the source code referred to that type. This helps make error messages easier to read for the user.

Null

The null type.

Int(IntTy)

An integer type.

UniversalInt

A universal integer type.

UnboundedInt

An unbounded integer type. This is the type integers have that are evaluated at compile time, e.g. as part of a range expression. Cannot be mapped to LLHD.

Enum(EnumTy)

An enumeration type.

Physical(PhysicalTy)

A physical type.

Access(Box<Ty>)

An access type.

Array(ArrayTy)

An array type.

File(Box<Ty>)

A file type.

Record(RecordTy)

A record type.

Subprog(SubprogTy)

A subprogram type.

Methods

impl Ty

pub fn kind_desc(&self) -> &'static str

Provide a textual description of the kind of type. For example, if called on an integer type, the result is "integer type", without any information on the exact nature of the integer.

pub fn is_int(&self) -> bool

Check if this type is an integer.

pub fn is_real(&self) -> bool

Check if this type is a real.