Struct moore_vhdl::ty2::PhysicalBasetype

pub struct PhysicalBasetype { /* fields omitted */ }

A physical base type.

In VHDL a physical type is an integer multiple of some measurement unit. A physical type has exactly one primary unit, and multiple secondary units defined as multiples of that primary unit.

Methods

impl PhysicalBasetype

pub fn new<I>(
range: Range<BigInt>,
units: I,
primary: usize
) -> PhysicalBasetype where
I: IntoIterator<Item = PhysicalUnit>,

Create a new physical type.

Example

use moore_vhdl::ty2::{PhysicalBasetype, PhysicalUnit, Range};
use moore_vhdl::common::name::get_name_table;

let ty = PhysicalBasetype::new(Range::ascending(0, 1_000_000), vec![
    PhysicalUnit::primary(get_name_table().intern("fs", false), 1),
    PhysicalUnit::secondary(get_name_table().intern("ps", false), 1_000, 1000, 0),
    PhysicalUnit::secondary(get_name_table().intern("ns", false), 1_000_000, 1000, 1),
], 0);

assert_eq!(format!("{}", ty), "0 to 1000000 units (fs, ps, ns)");