Struct moore_vhdl::ty2::UniversalRealType

pub struct UniversalRealType;

A universal real.

This is not strictly a separate type, but rather defined by the standard as the real type with the largest range. However since we can represent arbitrary numbers as f64, we use this special marker type.

Example

use moore_vhdl::ty2::{Type, UniversalRealType};

let ty = UniversalRealType;

assert_eq!(format!("{}", ty), "{universal real}");
assert_eq!(ty.is_scalar(), true);
assert_eq!(ty.is_discrete(), false);
assert_eq!(ty.is_numeric(), true);