Trait moore_vhdl::ty2::FloatingType

pub trait FloatingType: Type {
    fn as_type(&self) -> &dyn Type;

    fn range(&self) -> Option<&Range<f64>>;

    fn base_type(&self) -> &dyn Type;

    fn is_equal(&self, other: &dyn FloatingType) -> bool;

    fn resolution_func(&self) -> Option<usize> { ... }

    fn as_basetype(&self) -> Option<&FloatingBasetype> { ... }

    fn as_subtype(&self) -> Option<&FloatingSubtype> { ... }

    fn is_universal(&self) -> bool { ... }

    fn unwrap_basetype(&self) -> &FloatingBasetype { ... }

    fn unwrap_subtype(&self) -> &FloatingSubtype { ... }
}

A real type.

This can either be an FloatingBasetype or an FloatingSubtype.

Required methods

fn as_type(&self) -> &dyn Type

Convert to a type.

fn range(&self) -> Option<&Range<f64>>

The range of values this real can assume.

Universal integers return None here, as they do not have a value range associated with them.

fn base_type(&self) -> &dyn Type

The base type of this real.

fn is_equal(&self, other: &dyn FloatingType) -> bool

Check if two real types are equal.

Loading content...
Provided methods

fn resolution_func(&self) -> Option<usize>

The resolution function associated with this type.

fn as_basetype(&self) -> Option<&FloatingBasetype>

Returns Some if self is an FloatingBasetype, None otherwise.

fn as_subtype(&self) -> Option<&FloatingSubtype>

Returns Some if self is an FloatingSubtype, None otherwise.

fn is_universal(&self) -> bool

Checks whether this is a universal real type.

fn unwrap_basetype(&self) -> &FloatingBasetype

Returns an &FloatingBasetype or panics if the type is not a basetype.

fn unwrap_subtype(&self) -> &FloatingSubtype

Returns an &FloatingSubtype or panics if the type is not a subtype.

Loading content...